backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
aiosqlite
python-magic